requirements.txt
streamlit
pandas
openpyxl
requests
ollama
plotly
watchdog
psycopg2-binary==2.9.9
sqlalchemy==2.0.48
tabulate